testcases

+++ PluginKVS/include/util_t.h
@@ -82,6 +82,6 @@ struct doublehashfunc
 
 int nearest_power_two(int n);
 void create_integertestinput(std::string&,int,int,int,std::vector<int> &keys,std::vector<uint64_t> &ts);
-void create_timeseries_testinput(std::string&,int,int,std::vector<float>&,std::vector<uint64_t>&,std::vector<int>&);
+void create_timeseries_testinput(std::string&,int,int,std::vector<uint64_t>&,std::vector<float>&,std::vector<int>&);
 void create_dataworld_testinput(std::string&,int,int,std::vector<uint64_t>&,std::vector<uint64_t>&,std::vector<int>&);
 #endif
